// src/module_bindings/types.ts
// THIS FILE IS AUTOMATICALLY GENERATED BY SPACETIMEDB. EDITS TO THIS FILE
// WILL NOT BE SAVED. MODIFY TABLES IN YOUR MODULE SOURCE CODE INSTEAD.

/* eslint-disable */
/* tslint:disable */
import {
  TypeBuilder as __TypeBuilder,
  t as __t,
  type AlgebraicTypeType as __AlgebraicTypeType,
  type Infer as __Infer,
} from "spacetimedb";

export const Battle = __t.object("Battle", {
  from: __t.u8(),
  to: __t.u8(),
  attackerSeat: __t.u8(),
  defenderSeat: __t.i8(),
  attackerDice: __t.byteArray(),
  defenderDice: __t.byteArray(),
  attackerLosses: __t.u32(),
  defenderLosses: __t.u32(),
  rolls: __t.u32(),
  captured: __t.bool(),
});
export type Battle = __Infer<typeof Battle>;

export const Card = __t.object("Card", {
  id: __t.u64(),
  gameId: __t.u64(),
  territoryIdx: __t.i16(),
  symbol: __t.u8(),
  location: __t.string(),
  holderSeat: __t.i8(),
});
export type Card = __Infer<typeof Card>;

export const ChatMessage = __t.object("ChatMessage", {
  id: __t.u64(),
  gameId: __t.u64(),
  userId: __t.u64(),
  username: __t.string(),
  text: __t.string(),
  sentAt: __t.timestamp(),
});
export type ChatMessage = __Infer<typeof ChatMessage>;

export const Credential = __t.object("Credential", {
  userId: __t.u64(),
  salt: __t.string(),
  hash: __t.string(),
});
export type Credential = __Infer<typeof Credential>;

export const Game = __t.object("Game", {
  id: __t.u64(),
  name: __t.string(),
  creatorId: __t.u64(),
  status: __t.string(),
  createdAt: __t.timestamp(),
  startedAt: __t.option(__t.timestamp()),
  finishedAt: __t.option(__t.timestamp()),
  playerCount: __t.u8(),
  currentSeat: __t.u8(),
  phase: __t.string(),
  turnNumber: __t.u32(),
  reinforcements: __t.u32(),
  setsTraded: __t.u32(),
  conqueredThisTurn: __t.bool(),
  cardBonusTaken: __t.bool(),
  forcedTrade: __t.bool(),
  occupyFrom: __t.u8(),
  occupyTo: __t.u8(),
  get lastBattle() {
    return __t.option(Battle);
  },
  winnerId: __t.option(__t.u64()),
  winnerName: __t.string(),
});
export type Game = __Infer<typeof Game>;

export const GameLog = __t.object("GameLog", {
  id: __t.u64(),
  gameId: __t.u64(),
  text: __t.string(),
  at: __t.timestamp(),
});
export type GameLog = __Infer<typeof GameLog>;

export const GamePlayer = __t.object("GamePlayer", {
  id: __t.u64(),
  gameId: __t.u64(),
  userId: __t.u64(),
  username: __t.string(),
  joinedAt: __t.timestamp(),
  seat: __t.u8(),
  color: __t.string(),
  armiesToPlace: __t.u32(),
  cardCount: __t.u32(),
  eliminated: __t.bool(),
  forfeited: __t.bool(),
});
export type GamePlayer = __Infer<typeof GamePlayer>;

export const MyCards = __t.object("MyCards", {});
export type MyCards = __Infer<typeof MyCards>;

export const Presence = __t.object("Presence", {
  identity: __t.identity(),
  connections: __t.u32(),
});
export type Presence = __Infer<typeof Presence>;

export const Session = __t.object("Session", {
  identity: __t.identity(),
  userId: __t.u64(),
});
export type Session = __Infer<typeof Session>;

export const Territory = __t.object("Territory", {
  id: __t.u64(),
  gameId: __t.u64(),
  idx: __t.u8(),
  ownerSeat: __t.i8(),
  armies: __t.u32(),
});
export type Territory = __Infer<typeof Territory>;

export const User = __t.object("User", {
  id: __t.u64(),
  username: __t.string(),
  usernameLower: __t.string(),
  createdAt: __t.timestamp(),
  online: __t.bool(),
  gamesPlayed: __t.u32(),
  wins: __t.u32(),
});
export type User = __Infer<typeof User>;